switched calls from apiCalls.tsx to trackerApiCalls.tsx
