+++ src/components/JobApplications/NewJobApplication.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useUserLoggedContext } from '../../context/UserLoggedContext';
 import { statusMap, statusStyles} from "../JobApplicationUtilities";
-import { fetchContacts, fetchCompanies } from "../../apiCalls";
+import { fetchContacts, fetchCompanies } from "../../trackerApiCalls";
 import { postJobApplication } from '../../trackerApiCalls';
 
 interface Company {
